[FIX] - float

+++ math/0x03-probability/poisson.py
@@ -27,7 +27,7 @@ class Poisson:
                 raise TypeError('data must be a list')
             if len(data) < 2:
                 raise ValueError('data must contain multiple values')
-            self.lambtha = sum(data) / len(data)
+            self.lambtha = float(sum(data) / len(data))
 
     def pmf(self, k):
         """Calculates the value of the PMF
